Rename `aoai-simulated-api` to `aoai-api-simulator` (also snake-cased package name)
