+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## vNext
 
+- Migrate to current repo
+  - BREAKING: rename `aoai-simulated-api` to `aoai-api-simulator` in code (also for `aoai_simulated_api` package)
 - Return to sliding window rate limiting. This change moves from the limits package to a custom rate-limiting implementation to address performance with sliding windows (#20)
 - Update rate-limit handling for tokens based on experimentation (limited set of models currently - see #52)
