+++ loadtest/post_steps_added_latency.py
@@ -43,7 +43,7 @@ query_processor = QueryProcessor(
 print(f"metric_check_time: {metric_check_time}")
 check_results_query = """
 AppMetrics
-| where AppRoleName == "aoai-simulated-api" // only check the metrics for the simulated API
+| where AppRoleName == "aoai-api-simulator" // only check the metrics for the simulated API
 | summarize max(TimeGenerated)
 """
