+++ loadtest/post_steps_limits_requests.py
@@ -45,7 +45,7 @@ print(f"metric_check_time: {metric_check_time}")
 
 check_results_query = """
 AppMetrics
-| where AppRoleName == "aoai-simulated-api" // only check the metrics for the simulated API
+| where AppRoleName == "aoai-api-simulator" // only check the metrics for the simulated API
 | summarize max(TimeGenerated)
 """
