+++ sample.env
@@ -9,5 +9,5 @@ SIMULATOR_MODE=generate                                                         
 OPENAI_DEPLOYMENT_CONFIG_PATH=examples/openai_deployment_config.json            # The path to a JSON file that contains the deployment configuration
 
 #  Open Telemetry Config
-OTEL_SERVICE_NAME=aoai-simulated-api-local-dev                                  # The service name used by Open Telemetry
+OTEL_SERVICE_NAME=aoai-api-simulator-local-dev                                  # The service name used by Open Telemetry
 OTEL_METRIC_EXPORT_INTERVAL=10000                                               # Open Telemetry metric export interval (milliseconds)
