+++ src/aoai-api-simulator/.dockerignore
