+++ src/aoai-api-simulator/src/aoai_api_simulator/limiters.py
@@ -8,9 +8,9 @@ from typing import Awaitable, Callable
 
 from fastapi import Response
 
-from aoai_simulated_api import constants
-from aoai_simulated_api.metrics import simulator_metrics
-from aoai_simulated_api.models import Config, RequestContext
+from aoai_api_simulator import constants
+from aoai_api_simulator.metrics import simulator_metrics
+from aoai_api_simulator.models import Config, RequestContext
 
 logger = logging.getLogger(__name__)
