+++ src/aoai-api-simulator/src/aoai_api_simulator/main.py
@@ -5,8 +5,8 @@ from azure.monitor.opentelemetry import configure_azure_monitor
 
 # from opentelemetry import trace
 
-from aoai_simulated_api.config_loader import get_config_from_env_vars, set_config
-from aoai_simulated_api.app_builder import app as builder_app, apply_config
+from aoai_api_simulator.config_loader import get_config_from_env_vars, set_config
+from aoai_api_simulator.app_builder import app as builder_app, apply_config
 
 log_level = os.getenv("LOG_LEVEL") or "INFO"
