+++ tests/test_extension_load.py
@@ -2,9 +2,9 @@
 Test extension loading
 """
 
-from aoai_simulated_api.config_loader import initialize_config
-from aoai_simulated_api.generator.manager import get_default_generators
-from aoai_simulated_api.models import (
+from aoai_api_simulator.config_loader import initialize_config
+from aoai_api_simulator.generator.manager import get_default_generators
+from aoai_api_simulator.models import (
     Config,
     LatencyConfig,
     ChatCompletionLatency,
@@ -16,7 +16,7 @@ import requests
 
 from .test_uvicorn_server import UvicornTestServer
 
-from aoai_simulated_api.record_replay.handler import get_default_forwarders
+from aoai_api_simulator.record_replay.handler import get_default_forwarders
 
 API_KEY = "123456789"
